add gpx file to zip export
