+++ backend/automatic_walk_time_tables/utils/gpx_creator.py
@@ -4,9 +4,23 @@ import xml.etree.ElementTree as ET
 import re
 import gpxpy
 from gpxpy.gpx import GPX
+import requests
 
 from automatic_walk_time_tables.utils import path
 
+
+def fetch_data_for_uuid(uuid):
+    """
+    Fetches the data for the given UUID from the store API.
+    :param uuid: The UUID of the data to fetch
+    :return: The data for the given UUID or None if the data is not available
+    """
+
+    r = requests.post(os.environ["STORE_API_URL"] + "/retrieve", json={"uuid": uuid})
+    if r.status_code == 200:
+        return r.json()
+
+    return None
 
 def create_gpx_file(_path: path.Path, _way_points: path.Path):
     """
